IOS-6366 Refactoring
